requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.2
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
python-decouple==3.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.2
